refactor: Lift sequence out of the pulses module
