+++ src/qibolab/instruments/qm/controller.py
@@ -14,7 +14,8 @@ from qualang_tools.simulator_tools import create_simulator_controller_connection
 from qibolab.components import Channel, ChannelId, Config, DcChannel, IqChannel
 from qibolab.execution_parameters import ExecutionParameters
 from qibolab.instruments.abstract import Controller
-from qibolab.pulses import Delay, Pulse, PulseSequence, VirtualZ
+from qibolab.pulses import Delay, Pulse, VirtualZ
+from qibolab.sequence import PulseSequence
 from qibolab.sweeper import ParallelSweepers, Parameter
 from qibolab.unrolling import Bounds
