+++ src/qibolab/pulses/__init__.py
@@ -1,3 +1,2 @@
 from .envelope import *
-from .pulse import Delay, Pulse, VirtualZ
-from .sequence import PulseSequence
+from .pulse import Delay, Pulse, PulseLike, VirtualZ
