+++ src/qibolab/pulses/plot.py
@@ -6,10 +6,11 @@ from typing import Optional
 import matplotlib.pyplot as plt
 import numpy as np
 
+from qibolab.sequence import PulseSequence
+
 from .envelope import Waveform
 from .modulation import modulate
 from .pulse import Delay, Pulse, VirtualZ
-from .sequence import PulseSequence
 
 SAMPLING_RATE = 1
 """Default sampling rate in gigasamples per second (GSps).
