+++ src/qibolab/sweeper.py
@@ -60,7 +60,7 @@ class Sweeper:
             import numpy as np
             from qibolab.dummy import create_dummy
             from qibolab.sweeper import Sweeper, Parameter
-            from qibolab.pulses import PulseSequence
+            from qibolab.sequence import PulseSequence
             from qibolab import ExecutionParameters
